update: refactor bosyu directory
